Refactor Hearing into separate component and view
Split views/Hearing/index.js into separate view and component. This
way reusing the Hearing becomes easier and we can reuse the Hearing
component in other views for example.

This work makes admin UI implementation easier.
